Fix segmentation fault caused by std::string instance passed to memcpy
In the implementation of csmVector, when memory is reallocated,
the values in the previous pointer is copied directly using
memcpy(). This is not safe to use with std::string. Instances
created using placement new (as it is inside csmVector) must be
destroyed by calling their destructor, before the underlying
memory storage can be manipulated.

std::vector handles memory correctly, and switching to std::vector
resolves the segmentation fault.